Ensure VA-to-Offset maps only to file-backed segments
Corrects the Virtual Address to File Offset conversion logic in ELFHelper to use p_filesz instead of p_memsz. This prevents invalid mappings to non-file-backed sections (like .bss). Additionally, adds explicit error handling in the CLI to abort if a VA cannot be resolved, preventing reads from invalid file offsets

+++ src/CLI/main.cpp
@@ -61,7 +61,7 @@ int main(int argc, const char* argv[])
 	if (bIsVA)
 	{
 		offset = ELFHelper::GetFileOffset(file.data(), offset);
-		if (offset == -1)
+		if (offset == (uint64_t)-1)
 		{
 			std::cout << "Invalid Virtual Address or unable to convert to File Offset" << std::endl;
 			return 1;
